Updated to install latest docker
current docker release (1.10.3) requires kernel 4.x.
  add UEKR4 to yum repo list if docker set to 'true'
  only run docker script if docker set to 'true'
  only run desktop script if desktop set to 'true'
  only run update script if update set to 'true'

+++ script/docker.sh
@@ -1,17 +1,20 @@
 #!/bin/bash
 
-echo "==> Run the Docker installation script"
-curl -sSL https://get.docker.com | sh
+if [ $DOCKER != 0 ]
+then
+  echo "==> Run the Docker installation script"
+  curl -sSL https://get.docker.com | sh
 
-echo "==> Create the docker group"
-# Add the docker group if it doesn't already exist
-groupadd docker
+  echo "==> Create the docker group"
+  # Add the docker group if it doesn't already exist
+  groupadd docker
 
-echo "==> Add the connected "${USER}" to the docker group."
-gpasswd -a ${USER} docker
-gpasswd -a ${SSH_USERNAME} docker
+  echo "==> Add the connected "${USER}" to the docker group."
+  gpasswd -a ${USER} docker
+  gpasswd -a ${SSH_USERNAME} docker
 
-echo "==> Starting docker"
-service docker start
-echo "==> Enabling docker to start on reboot"
-chkconfig docker on
+  echo "==> Starting docker"
+  service docker start
+  echo "==> Enabling docker to start on reboot"
+  chkconfig docker on
+fi
